fix: address pr review feedback for scheduled flag changes ui
