+++ ui/dashboard/src/pages/feature-flag-details/elements/scheduled-changes-banner/utils.ts
@@ -1,0 +1,9 @@
+export const formatScheduledDate = (timestamp: string | number): string => {
+  const date = new Date(Number(timestamp) * 1000);
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  return `${year}/${month}/${day} ${hours}:${minutes}`;
+};
