+++ ui/dashboard/src/pages/feature-flag-details/elements/scheduled-changes-panel/schedule-card.tsx
@@ -19,21 +19,12 @@ import Icon from 'components/icon';
 import { Popover, PopoverOption, PopoverValue } from 'components/popover';
 import { Tooltip } from 'components/tooltip';
 import DateTooltip from 'elements/date-tooltip';
+import { formatScheduledDate } from '../scheduled-changes-banner/utils';
 
 export type ScheduleCardAction =
   | 'APPLY_NOW'
   | 'EDIT_SCHEDULE'
   | 'CANCEL_SCHEDULE';
-
-const formatScheduledDate = (timestamp: string | number): string => {
-  const date = new Date(Number(timestamp) * 1000);
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-  return `${year}/${month}/${day} ${hours}:${minutes}`;
-};
 
 const STATUS_STYLE: Record<
   string,
